Add a license
Eventually, we would make this code available. It is a good practice to always add a license in the very initial phase of repo creation. This ensures that if anyone wants to use a older version of the code, they still would need to `disclose source` and `contribute` back to upstream. I hope this makes sense.